dev: move config.js to repo root
