+++ server/models.js
@@ -1,9 +1,10 @@
 /* eslint-disable global-require */
+import path from 'path';
 import Sequelize from 'sequelize';
 import { createIncludeUserModel } from './utils/queryHelpers/includeUserModel';
 
 if (process.env.NODE_ENV !== 'production' && process.env.NODE_ENV !== 'test') {
-	require('./config.js');
+	require(path.join(process.cwd(), 'config.js'));
 }
 
 const useSSL = process.env.DATABASE_URL.indexOf('localhost') === -1;
