+++ tools/dashboardMigrations/findBrokenFirebase.js
@@ -2,7 +2,7 @@
 import { Branch } from '../../server/models';
 
 const firebaseAdmin = require('firebase-admin');
-require('../../server/config');
+require('../../config');
 
 const serviceAccount = JSON.parse(
 	Buffer.from(process.env.FIREBASE_SERVICE_ACCOUNT_BASE64, 'base64').toString(),
